pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kubecuro"
version = "0.1.0"
description = "Enterprise-grade Kubernetes YAML healing and auditing tool."
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Apache-2.0"}
authors = [
    {name = "Nishar A Sunkesala", email = "fixmyk8s@gmail.com"}
]
keywords = ["kubernetes", "yaml", "devops", "sre", "automation"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    
]

# External Dependencies
dependencies = [
    "ruamel.yaml>=0.17.21",
    "rich>=12.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "mypy>=1.0",
]

# --- THE CORRECTED ENTRY POINT ---
# Maps to: src/kubecuro/cli/main.py -> function main()
[project.scripts]
kubecuro = "kubecuro.cli.main:main"

[project.urls]
Homepage = "https://github.com/fixmyk8s/kubecuro"
Issues = "https://github.com/fixmyk8s/kubecuro/issues"
Repository = "https://github.com/fixmyk8s/kubecuro.git"

# --- SETUPTOOLS CONFIGURATION ---
[tool.setuptools]
# Tells setuptools that 'src' is the root of your code
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
# Automatically find kubecuro, kubecuro.cli, kubecuro.core, etc.
where = ["src"]
exclude = ["tests*", "catalog*"]

[tool.setuptools.package-data]
# Ensures your distilled JSON maps are included in the package
kubecuro = ["assets/*"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
